requirements.txt

torchio==0.18.84
scipy==1.9.0
numpy==1.23.2
torchsummary==1.5.1
